backend/requirements.txt
fastapi
uvicorn[standard]
requests
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
httpx
langgraph
langchain_community
APScheduler
agent